+++ src/pczt/verify.rs
@@ -4,7 +4,7 @@ use crate::{
     keys::{FullViewingKey, SpendValidatingKey},
     note::{ExtractedNoteCommitment, Rho},
     value::ValueCommitment,
-    Note, NoteVersion,
+    Note,
 };
 
 impl super::Bundle {
@@ -111,7 +111,7 @@ impl super::Spend {
             self.value.ok_or(VerifyError::MissingValue)?,
             self.rho.ok_or(VerifyError::MissingRho)?,
             self.rseed.ok_or(VerifyError::MissingRandomSeed)?,
-            NoteVersion::DEFAULT,
+            self.note_version,
         )
         .into_option()
         .ok_or(VerifyError::InvalidSpendNote)?;
@@ -170,7 +170,7 @@ impl super::Output {
             self.value.ok_or(VerifyError::MissingValue)?,
             Rho::from_nf_old(spend.nullifier),
             self.rseed.ok_or(VerifyError::MissingRandomSeed)?,
-            NoteVersion::DEFAULT,
+            self.note_version,
         )
         .into_option()
         .ok_or(VerifyError::InvalidOutputNote)?;